requirements.txt
aniso8601==9.0.1
attrs==24.2.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-MySQL==1.5.2
Flask-MySQLdb==2.0.0
flask-restx==1.3.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mysqlclient==2.2.5
PyJWT==2.9.0
PyMySQL==1.1.1
pytz==2024.2
referencing==0.35.1
rpds-py==0.21.0
Werkzeug==3.1.2
